Attempt to fix issue 376
Adding a parameter map to allow implementation defined casting and introducing an error, if the map contains illegal values.
My first move was to use a `map(xs:QName, item()*)` and to say that the keys must be in an (non-XProc)-namespaces. But as we tend to use untyped maps anywhere, it has be to `map(*)` here also.